+++ src/app/cart/cart.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CartRoutingModule, cartRouterComponents } from './cart-routing.module';
 import { OrderFormComponent } from './components/order-form/order-form.component';
 import { CartComponent } from './cart.component';
@@ -20,6 +20,7 @@ import { CartComponent } from './cart.component';
     CommonModule,
     SharedModule,
     FormsModule,
+    ReactiveFormsModule,
     CartRoutingModule
   ],
   exports: [