+++ changelog.md
@@ -1,3 +1,1 @@
-1. Added store (including effects and entities) for products feature area. All usings of products-service were removed from components.
-2. Added router-store. All direct usings of Router removed from products feature area and replaced with store router actions.
-3. Added  products preloading guard, which uses appSettings guard inside, that lets us to load appSettings before the loading of products components. 
+1. Added reactive form with validations, form arrays and form groups (see src\app\cart\components\order-form\)